+++ lms/loan.py
@@ -13,6 +13,7 @@ from lxml import etree
 from utils.responder import respondWithFailure, respondWithSuccess
 
 import lms
+from lms.firebase import FirebaseAdmin
 from lms.lms.doctype.approved_terms_and_conditions.approved_terms_and_conditions import (
     ApprovedTermsandConditions,
 )
@@ -307,6 +308,15 @@ def esign_done(**kwargs):
                 from frappe.core.doctype.sms_settings.sms_settings import send_sms
 
                 frappe.enqueue(method=send_sms, receiver_list=receiver_list, msg=msg)
+
+                fcm_notification = frappe.get_doc(
+                    "Spark Push Notification",
+                    "Topup E-signing was successful",
+                    fields=["*"],
+                )
+                lms.send_spark_push_notification(
+                    fcm_notification=fcm_notification, customer=customer
+                )
 
             return utils.respondWithSuccess()
         except requests.RequestException as e:
@@ -1134,6 +1144,12 @@ def loan_details(**kwargs):
         loan_margin_shortfall = loan.get_margin_shortfall()
         if loan_margin_shortfall.get("__islocal", None):
             loan_margin_shortfall = None
+
+        res["is_sell_triggered"] = 0
+        if loan_margin_shortfall:
+            if loan_margin_shortfall.status == "Sell Triggered":
+                res["is_sell_triggered"] = 1
+
         unpledge_application_exist = frappe.get_all(
             "Unpledge Application",
             filters={"loan": loan.name, "status": "Pending"},
@@ -1154,6 +1170,11 @@ def loan_details(**kwargs):
             )
 
         res["amount_available_for_withdrawal"] = loan.maximum_withdrawable_amount()
+
+        # Pledgor boid of particular loan
+        res["pledgor_boid"] = frappe.db.get_value(
+            "Collateral Ledger", {"loan": loan.name}, "pledgor_boid"
+        )
 
         return utils.respondWithSuccess(data=res)
     except utils.exceptions.APIException as e:
@@ -1448,6 +1469,18 @@ def loan_payment(**kwargs):
             from frappe.core.doctype.sms_settings.sms_settings import send_sms
 
             frappe.enqueue(method=send_sms, receiver_list=receiver_list, msg=msg)
+
+            fcm_notification = frappe.get_doc(
+                "Spark Push Notification", "Payment failed", fields=["*"]
+            )
+            lms.send_spark_push_notification(
+                fcm_notification=fcm_notification,
+                message=fcm_notification.message.format(
+                    amount=data.get("amount"), loan=loan.name
+                ),
+                loan=loan.name,
+                customer=customer,
+            )
             return utils.respondWithSuccess(message="Check Loan Payment Log.")
 
         if data.get("loan_margin_shortfall_name", None) and not data.get("is_failed"):
@@ -1487,6 +1520,11 @@ def loan_payment(**kwargs):
             #             loan.name
             #         ),
             #     )
+            if loan_margin_shortfall.status == "Sell Triggered":
+                return utils.respondWithFailure(
+                    status=417,
+                    message=frappe._("Sale is Triggered"),
+                )
             if loan_margin_shortfall.status == "Pending":
                 loan_margin_shortfall.status = "Request Pending"
                 loan_margin_shortfall.save(ignore_permissions=True)
@@ -1496,6 +1534,14 @@ def loan_payment(**kwargs):
                 "Notification", "Margin Shortfall Action Taken", method="send", doc=doc
             )
             msg = "Dear Customer,\nThank you for taking action against the margin shortfall.\nYou can view the 'Action Taken' summary on the dashboard of the app under margin shortfall banner. Spark Loans"
+            fcm_notification = frappe.get_doc(
+                "Spark Push Notification",
+                "Margin shortfall – Action taken",
+                fields=["*"],
+            )
+            lms.send_spark_push_notification(
+                fcm_notification=fcm_notification, loan=loan.name, customer=customer
+            )
             # receiver_list = list(
             #     set([str(customer.phone), str(customer.get_kyc().mobile_number)])
             # )
@@ -2239,14 +2285,14 @@ def validate_securities_for_unpledge(securities, loan):
     if not securities or (
         type(securities) is not dict and "list" not in securities.keys()
     ):
-        raise utils.ValidationException(
+        raise utils.exceptions.ValidationException(
             {"securities": {"required": frappe._("Securities required.")}}
         )
 
     securities = securities["list"]
 
     if len(securities) == 0:
-        raise utils.ValidationException(
+        raise utils.exceptions.ValidationException(
             {"securities": {"required": frappe._("Securities required.")}}
         )
 
@@ -2323,7 +2369,9 @@ def validate_securities_for_unpledge(securities, loan):
                 break
 
     if not securities_valid:
-        raise utils.ValidationException({"securities": {"required": message}})
+        raise utils.exceptions.ValidationException(
+            {"securities": {"required": message}}
+        )
 
     return securities
 
@@ -2390,7 +2438,7 @@ def loan_unpledge_request(**kwargs):
         )
 
         if token.expiry <= frappe.utils.now_datetime():
-            return utils.respondUnauthorized(message=frappe._("Pledge OTP Expired."))
+            return utils.respondUnauthorized(message=frappe._("Unpledge OTP Expired."))
 
         frappe.db.begin()
 
@@ -2542,6 +2590,11 @@ def sell_collateral_request(**kwargs):
             loan_margin_shortfall = frappe.get_doc(
                 "Loan Margin Shortfall", data.get("loan_margin_shortfall_name")
             )
+            if loan_margin_shortfall.status == "Sell Triggered":
+                return utils.respondWithFailure(
+                    status=417,
+                    message=frappe._("Sale is Triggered"),
+                )
             pending_sell_collateral_application = frappe.get_all(
                 "Sell Collateral Application",
                 filters={
@@ -2573,6 +2626,14 @@ def sell_collateral_request(**kwargs):
                 "Notification", "Margin Shortfall Action Taken", method="send", doc=doc
             )
             msg = "Dear Customer,\nThank you for taking action against the margin shortfall.\nYou can view the 'Action Taken' summary on the dashboard of the app under margin shortfall banner. Spark Loans"
+            fcm_notification = frappe.get_doc(
+                "Spark Push Notification",
+                "Margin shortfall – Action taken",
+                fields=["*"],
+            )
+            lms.send_spark_push_notification(
+                fcm_notification=fcm_notification, loan=loan.name, customer=customer
+            )
 
         sell_collateral_application.insert(ignore_permissions=True)
 
@@ -2603,14 +2664,14 @@ def validate_securities_for_sell_collateral(securities, loan_name):
     if not securities or (
         type(securities) is not dict and "list" not in securities.keys()
     ):
-        raise utils.ValidationException(
+        raise utils.exceptions.ValidationException(
             {"securities": {"required": frappe._("Securities required.")}}
         )
 
     securities = securities["list"]
 
     if len(securities) == 0:
-        raise utils.ValidationException(
+        raise utils.exceptions.ValidationException(
             {"securities": {"required": frappe._("Securities required.")}}
         )
 
@@ -2660,7 +2721,9 @@ def validate_securities_for_sell_collateral(securities, loan_name):
                 break
 
     if not securities_valid:
-        raise utils.ValidationException({"securities": {"required": message}})
+        raise utils.exceptions.ValidationException(
+            {"securities": {"required": message}}
+        )
 
     return securities